# docs/openapi.planner.yaml
openapi: 3.1.0
info:
  title: Guia de Milhas Planner API
  version: 2026-02-11
  description: Contrato formal do endpoint de geração de planner.
servers:
  - url: https://guiademilhas.app
    description: Produção
  - url: http://localhost:3000
    description: Local
paths:
  /api/planner/generate:
    post:
      summary: Gera relatório de planejamento com milhas
      operationId: postPlannerGenerate
      tags:
        - Planner
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlannerGenerateRequest'
      responses:
        '200':
          description: Relatório gerado com sucesso
          headers:
            x-request-id:
              description: Correlation id da requisição
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlannerGenerateSuccess'
        '400':
          description: Payload inválido
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/PlannerProblem'
        '401':
          description: Não autenticado
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/PlannerProblem'
        '429':
          description: Rate limit excedido
          headers:
            Retry-After:
              description: Tempo em segundos até nova tentativa
              schema:
                type: integer
                minimum: 1
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/PlannerProblem'
        '500':
          description: Erro interno
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/PlannerProblem'
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  schemas:
    PlannerGenerateRequest:
      type: object
      additionalProperties: false
      required: [preferences]
      properties:
        locale:
          type: string
          description: >
            Opcional. O backend normaliza para locale suportado (ex.: pt-BR/en).
            Valores fora do conjunto esperado caem em fallback seguro.
        source:
          type: string
          enum: [landing_planner]
          description: Opcional. Origem de funil para mensuração de conversão.
        preferences:
          type: object
          additionalProperties: false
          required:
            - data_ida
            - data_volta
            - flex_dias
            - origens
            - destinos
            - num_adultos
            - num_chd
            - num_inf
            - idades_chd_inf
            - preferencia_voo
            - horarios_voo
            - bagagem
            - programas_milhas
            - programas_bancos
            - vistos_existentes
            - orcamento_brl
            - tolerancia_risco
            - perfil
            - hospedagem_padrao
            - bairros_pref
            - restricoes
          properties:
            data_ida:
              type: string
              pattern: '^\d{4}-\d{2}-\d{2}$'
            data_volta:
              type: string
              pattern: '^\d{4}-\d{2}-\d{2}$'
            flex_dias:
              type: string
              pattern: '^(?:[0-9]|[12][0-9]|30)$'
            origens:
              type: string
              minLength: 2
              maxLength: 200
            destinos:
              type: string
              minLength: 2
              maxLength: 200
            num_adultos:
              oneOf:
                - type: integer
                  minimum: 1
                  maximum: 9
                - type: string
                  pattern: '^[1-9]$'
              description: Backend aceita inteiro ou string numerica (coercao segura).
            num_chd:
              oneOf:
                - type: integer
                  minimum: 0
                  maximum: 9
                - type: string
                  pattern: '^[0-9]$'
              description: Backend aceita inteiro ou string numerica (coercao segura).
            num_inf:
              oneOf:
                - type: integer
                  minimum: 0
                  maximum: 9
                - type: string
                  pattern: '^[0-9]$'
              description: Backend aceita inteiro ou string numerica (coercao segura).
            idades_chd_inf:
              type: string
              maxLength: 120
            preferencia_voo:
              type: string
              enum: [direto, 1_conexao, indiferente]
            horarios_voo:
              type: string
              enum: [qualquer, manha, tarde, noite, madrugada, evitar_madrugada]
            bagagem:
              type: string
              enum: [mao, 1_despachada, mais_despachadas]
            programas_milhas:
              type: string
              maxLength: 240
            programas_bancos:
              type: string
              maxLength: 240
            vistos_existentes:
              type: string
              maxLength: 240
            orcamento_brl:
              type: string
              maxLength: 120
            tolerancia_risco:
              type: string
              enum: [baixa, media, alta]
            perfil:
              type: string
              maxLength: 240
            hospedagem_padrao:
              type: string
              enum: ['3', '4', '5', indiferente]
            bairros_pref:
              type: string
              maxLength: 240
            restricoes:
              type: string
              maxLength: 800
    PlannerGenerateSuccess:
      type: object
      additionalProperties: false
      required: [schemaVersion, generatedAt, report, mode]
      properties:
        schemaVersion:
          type: string
          example: '2026-02-11'
        generatedAt:
          type: string
          format: date-time
        mode:
          type: string
          enum: [ai, fallback]
        report:
          $ref: '#/components/schemas/PlannerReport'
    PlannerReport:
      type: object
      additionalProperties: false
      required: [title, summary, sections, assumptions]
      properties:
        title:
          type: string
        summary:
          type: string
        sections:
          type: array
          minItems: 4
          maxItems: 8
          items:
            $ref: '#/components/schemas/PlannerSection'
        assumptions:
          type: array
          maxItems: 10
          items:
            type: string
    PlannerSection:
      type: object
      additionalProperties: false
      required: [title, items]
      properties:
        title:
          type: string
        items:
          type: array
          minItems: 2
          maxItems: 6
          items:
            type: string
    PlannerProblem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
        code:
          type: string
          enum:
            - planner_invalid_request
            - planner_unauthorized
            - planner_rate_limited
            - planner_internal_error
        retryAfterSeconds:
          type: integer
          minimum: 1
        error:
          type: string
    ApiError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
        requestId:
          type: string